+++ catalog/pages/buttons/index.md
@@ -15,10 +15,6 @@ rows:
     Type: string
     Default: '`regular`'
     Notes: Determines button size. Possible sizes are `small`, `regular`, `large`
-  - Prop: isLoading
-    Type: bool
-    Default: '`false`'
-    Notes: Determines the button loading state. Possible states are `true`, `false`
   - Prop: href
     Type: string
     Notes: Determines wether underlying HTML element should be an `a`
@@ -49,9 +45,6 @@ rows:
                     <Button size="large">My cool button</Button>
                 </Spacing>
                 <Spacing top={{small: "moderate"}}>
-                    <Button size="large" isLoading>My cool button</Button>
-                </Spacing>
-                <Spacing top={{small: "moderate"}}>
                     <Button size="large" disabled>My cool button</Button>
                 </Spacing>
             </Column>
@@ -75,9 +68,6 @@ rows:
                 </Spacing>
                 <Spacing top={{small: "moderate"}}>
                     <Button variant="special" size="large">My cool button</Button>
-                </Spacing>
-                <Spacing top={{small: "moderate"}}>
-                    <Button variant="special" size="large" isLoading>My cool button</Button>
                 </Spacing>
                 <Spacing top={{small: "moderate"}}>
                     <Button variant="special" size="large" disabled>My cool button</Button>
@@ -105,9 +95,6 @@ rows:
                     <Button variant="outline" size="large">My cool button</Button>
                 </Spacing>
                 <Spacing top={{small: "moderate"}}>
-                    <Button variant="outline" size="large" isLoading>My cool button</Button>
-                </Spacing>
-                <Spacing top={{small: "moderate"}}>
                     <Button variant="outline" size="large" disabled>My cool button</Button>
                 </Spacing>
             </Column>
@@ -131,9 +118,6 @@ rows:
                 </Spacing>
                 <Spacing top={{small: "moderate"}}>
                     <Button variant="outlineGray" size="large">My cool button</Button>
-                </Spacing>
-                <Spacing top={{small: "moderate"}}>
-                    <Button variant="outlineGray" size="large" isLoading>My cool button</Button>
                 </Spacing>
                 <Spacing top={{small: "moderate"}}>
                     <Button variant="outlineGray" size="large" disabled>My cool button</Button>
@@ -161,9 +145,6 @@ rows:
                     <Button variant="transparent" size="large">My cool button</Button>
                 </Spacing>
                 <Spacing top={{small: "moderate"}}>
-                    <Button variant="transparent" size="large" isLoading>My cool button</Button>
-                </Spacing>
-                <Spacing top={{small: "moderate"}}>
                     <Button variant="transparent" size="large" disabled>My cool button</Button>
                 </Spacing>
             </Column>
@@ -187,6 +168,50 @@ rows:
                 </Spacing>
                 <Spacing top={{small: "moderate"}}>
                     <Button size="large" href="#">My cool button</Button>
+                </Spacing>
+            </Column>
+        </Row>
+    </Container>
+</ThemeProvider>
+```
+
+### ButtonWithLoading
+
+Since this component is an extended version of Button, it has the same props except one additional prop.
+
+### Props
+
+```table
+span: 6
+rows:
+  - Prop: isLoading
+    Type: bool
+    Default: '`false`'
+    Notes: Determines the button loading state. Possible states are `true`, `false`
+```
+
+```react
+---
+<ThemeProvider theme={{ themeName: 'tm' }}>
+    <Container>
+        <Row>
+            <Column small={3}/>
+            <Column small={6}>
+                <ButtonWithLoading>My cool button</ButtonWithLoading>
+                <Spacing top={{small: "cozy"}}>
+                    <ButtonWithLoading isLoading>My cool button</ButtonWithLoading>
+                </Spacing>
+                <Spacing top={{small: "cozy"}}>
+                    <ButtonWithLoading variant="outline">My cool button</ButtonWithLoading>
+                </Spacing>
+                <Spacing top={{small: "cozy"}}>
+                    <ButtonWithLoading variant="outline" isLoading>My cool button</ButtonWithLoading>
+                </Spacing>
+                <Spacing top={{small: "cozy"}}>
+                    <ButtonWithLoading variant="transparent">My cool button</ButtonWithLoading>
+                </Spacing>
+                <Spacing top={{small: "cozy"}}>
+                    <ButtonWithLoading variant="transparent" isLoading>My cool button</ButtonWithLoading>
                 </Spacing>
             </Column>
         </Row>
