+++ src/components/ButtonWithLoading/Loader/__tests__/index.spec.js
@@ -1,0 +1,17 @@
+import React from "react";
+import renderer from "react-test-renderer";
+
+import Loader from "..";
+
+describe("Loader", () => {
+  it("should match snapshot", () => {
+    const props = {
+      children: "Children"
+    };
+
+    const component = renderer.create(<Loader {...props} />);
+    const tree = component.toJSON();
+
+    expect(tree).toMatchSnapshot();
+  });
+});
